pyproject.toml
[project]
name = "rag-from-scratch"
version = "0.1.0"
description = "rag-from-scratch"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
"bs4",
"ipython",
"ipykernel",
"langchain",
"langchain-openai",
"langchain-anthropic",
"langchain_community",
"tiktoken",
"chromadb",
"langchainhub",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"